refactor(common/lmlayer): move ASCII word breaker to new package
