+++ common/lexical-model-word-breakers/.gitignore
@@ -1,0 +1,3 @@
+*.js
+*.js.map
+*.d.ts
