+++ common/lexical-model-word-breakers/src/index.ts
@@ -1,0 +1,2 @@
+import {ascii} from "./ascii-word-breaker";
+export {ascii};
