+++ common/lexical-model-word-breakers/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "declaration": true,
+    "sourceMap": true,
+    "outDir": "lib",
+    "allowJs": false,
+    "lib": ["es6"],
+    "target": "es5"
+  },
+  "include": [
+	"src/**/*"
+  ],
+  "exclude": [
+	"node_modules",
+	"test/**/*.ts"
+  ]
+}
